6. Setup notifications, schedules. Refactor after review
